+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @buildpad/hooks
 
+## 1.6.0
+
+### Patch Changes
+
+- @buildpad/services@1.6.0
+- @buildpad/types@1.6.0
+- @buildpad/utils@1.6.0
+
 ## 1.5.0
 
 ### Minor Changes